configure MovieCard stories
