+++ client/src/components/MovieCard.tsx
@@ -24,7 +24,7 @@ const IMAGE_PASS: string = 'https://image.tmdb.org/t/p/original';
 const MovieCard: React.FC<MovieCardPropsInterface> = (props) => {
   const cutOverview = (input: string, maxChars: number) =>
     input.length <= maxChars ? input : input.slice(0, maxChars) + '...';
-
+  console.log(props.movie);
   return (
     <Card style={cardStyle}>
       <CardHeader
