+++ client/tsconfig.json
@@ -9,6 +9,8 @@
     "strict": true,
     "forceConsistentCasingInFileNames": true,
     "noFallthroughCasesInSwitch": true,
+    "strictNullChecks": true,
+    "strictFunctionTypes": true,
     "module": "esnext",
     "moduleResolution": "node",
     "resolveJsonModule": true,
